+++ datanode/api/config.go
@@ -38,6 +38,7 @@ type Config struct {
 	CoreNodeGRPCPort         int               `long:"core-node-grpc-port"`
 	RateLimit                ratelimit.Config  `group:"rate-limits"`
 	MaxSubscriptionPerClient uint32            `long:"max-subscription-per-client"`
+	TrustedProxies           []string          `long:"trusted-proxy"`
 }
 
 // NewDefaultConfig creates an instance of the package specific configuration, given a
@@ -57,5 +58,6 @@ func NewDefaultConfig() Config {
 		CoreNodeGRPCPort:         3002,
 		RateLimit:                ratelimit.NewDefaultConfig(),
 		MaxSubscriptionPerClient: 250,
+		TrustedProxies:           []string{},
 	}
 }